+++ Bicimad.Core/DomainObjects/UserHistory.cs
@@ -31,7 +31,7 @@ namespace Bicimad.Core.DomainObjects
         public bool Finished { get; set; }
 
         public string TotalDiscount { get; set; }
-        public int FinalPrice { get; set; }
+        public double FinalPrice { get; set; }
 
         [Key, StringLength(13)]
         public string Id { get; set; }
